src/app/setting/setting.component.spec.ts
import { async, ComponentFixture, TestBed } from '@angular/core/testing';

import { SettingComponent } from './setting.component';
import { FormsModule, ReactiveFormsModule } from '@angular/forms';
import { MatInputModule, MatButtonModule } from '@angular/material';
import { NgxsModule } from '@ngxs/store';
import { BrowserAnimationsModule } from '@angular/platform-browser/animations';

describe('SettingComponent', () => {
  let component: SettingComponent;
  let fixture: ComponentFixture<SettingComponent>;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [ SettingComponent ],
      imports: [
        BrowserAnimationsModule,
        FormsModule,
        ReactiveFormsModule,
        MatInputModule,
        MatButtonModule,
        NgxsModule.forRoot([])]
    })
    .compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(SettingComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should be created', () => {
    expect(component).toBeTruthy();
  });

  describe('Render', () => {
    describe('when loaded', () => {
      it('then have three input fields', () => {
        const settingComponent: HTMLElement = fixture.nativeElement;
        const formElements = settingComponent.querySelectorAll("mat-form-field");
        expect(formElements.length).toEqual(3)
      })

      it('then have a save button disabled', () => {
        const settingComponent: HTMLElement = fixture.nativeElement;
        const buttonElement = settingComponent.querySelector(".setting button:disabled");        
        expect(buttonElement).toBeTruthy()
      })
    })
  })
});
